feat: mongoDB connection logic
