+++ package.json
@@ -14,6 +14,7 @@
   "dependencies": {
     "ejs": "^3.1.10",
     "express": "^4.21.2",
+    "mongodb": "^3.2.6",
     "nodemon": "^3.1.9"
   }
 }
